Refactor mint chart data getter
